+++ tests/ciscotest.rb
@@ -137,6 +137,10 @@ class CiscoTestCase < TestCase
     flunk(message)
   end
 
+  def validate_property_excluded?(feature, property)
+    !node.cmd_ref.supports?(feature, property)
+  end
+
   def interfaces
     unless @@interfaces
       # Build the platform_info, used for interface lookup
@@ -149,7 +153,7 @@ class CiscoTestCase < TestCase
       end
       # rubocop:enable Style/ClassVars
     end
-    abort "No suitable interfaces found on #{node} for this test" if
+    skip "No suitable interfaces found on #{node} for this test" if
       @@interfaces.empty?
     @@interfaces
   end
@@ -203,5 +207,18 @@ class CiscoTestCase < TestCase
     else
       ["default interface #{intf_name}"]
     end
+  end
+
+  def mt_full_interface?
+    # MT-full tests require a specific linecard; either because they need a
+    # compatible interface or simply to enable the features. Either way
+    # we will provide an appropriate interface name if the linecard is present.
+    # Example 'show mod' output to match against:
+    #   '9  12  10/40 Gbps Ethernet Module  N7K-F312FQ-25 ok'
+    sh_mod = @device.cmd("sh mod | i '^[0-9]+.*N7K-F3'")[/^(\d+)\s.*N7K-F3/]
+    slot = sh_mod.nil? ? nil : Regexp.last_match[1]
+    skip('Unable to find compatible interface in chassis') if slot.nil?
+
+    "ethernet#{slot}/1"
   end
 end