+++ tests/test_evpn_vni.rb
@@ -28,8 +28,16 @@ class TestEvpnVni < CiscoTestCase
     # ensure we are starting with a clean slate for each test.
     super
     config('no feature bgp')
-    config('no nv overlay evpn')
-    config('no evpn')
+
+    # Some platforms complain when nv overlay is not configured
+    config_no_warn('no nv overlay evpn')
+
+    # Some platforms remove the 'evpn' command when 'no nv overlay evpn'
+    # is processed, while others must remove it explicitly.
+    config_no_warn('no evpn')
+
+  rescue RuntimeError => e
+    hardware_supports_feature?(e.message)
   end
 
   def test_create_and_destroy
@@ -40,6 +48,9 @@ class TestEvpnVni < CiscoTestCase
     vni.destroy
     vni_list = EvpnVni.vnis
     refute(vni_list.key?('4096'), 'Error: failed to destroy evpn vni 4096')
+
+  rescue RuntimeError => e
+    hardware_supports_feature?(e.message)
   end
 
   def test_vni_collection
@@ -59,6 +70,9 @@ class TestEvpnVni < CiscoTestCase
     assert_empty(vni.route_distinguisher,
                  'vni route_distinguisher should *NOT* be configured')
     vni.destroy
+
+  rescue RuntimeError => e
+    hardware_supports_feature?(e.message)
   end
 
   # test route_target
@@ -66,7 +80,11 @@ class TestEvpnVni < CiscoTestCase
     vni = EvpnVni.new(4096)
 
     # test route target both auto and route target both auto evpn
-    opts = [:both, :import, :export]
+    if Utils.nexus_i2_image
+      opts = [:import, :export]
+    else
+      opts = [:both, :import, :export]
+    end
 
     # Master list of communities to test against
     master = ['1.2.3.4:55', '2:2', '55:33', 'auto']
@@ -88,6 +106,9 @@ class TestEvpnVni < CiscoTestCase
     route_target_tester(vni, opts, should, 'Test 4')
 
     vni .destroy
+
+  rescue RuntimeError => e
+    hardware_supports_feature?(e.message)
   end
 
   def route_target_tester(vni, opts, should, test_id)